Split ast_tests into per-domain sibling modules (#73)
`tests/ast_tests.rs` was already 632 lines on main, over the 400-line
cap in AGENTS.md, and this branch's new cases pushed it further out.
Turn the file into an aggregator following the `#[path = "..."] mod x;`
convention used by `tests/logging_stderr_tests.rs`, and move the cases
into `tests/ast_tests/` grouped by manifest domain:

- `support.rs`: the shared `parse_manifest` wrapper
- `parsing.rs`: required and optional fields, unknown or ill-shaped
  sections, target flags
- `string_or_list.rs`: the scalar-or-sequence shape plus the
  `map_each`, `to_string_vec` and `as_single` case tables
- `macros.rs`: macro parsing and serialization round-trips
- `actions.rs`: the `actions:` phony and always defaults
- `manifest_files.rs`: the `manifest::from_path` cases

Every resulting file is under 400 lines and all assertions are
preserved. The `to_paths` case table is dropped because `to_paths` is
once again private to the IR module; `tests/ir_from_manifest_tests.rs`
already exercises scalar, list and absent path fields through the public
lowering pipeline, and the per-item traversal it delegates to is covered
directly by the `map_each` cases. Add assertion messages to the bare
`assert!` calls that the move touched, and an empty-list `map_each`
case.
